Simplification des packages
