+++ Revision_Exams/Intra_2/Exercice_1/graphiques/CercleGraphique.java
@@ -1,9 +1,9 @@
-package Librairies.Graphismes.src.com.moumene.graphiques;
+package Revision_Exams.Intra_2.Exercice_1.graphiques;
 
 import java.awt.Color;
 import java.awt.Graphics;
 
-import Librairies.Graphismes.src.com.moumene.composants.Cercle;
+import Revision_Exams.Intra_2.Exercice_1.composants.Cercle;
 
 /**
  * Represente l'objet graphique Cercle
