+++ Revision_Exams/Intra_2/Exercice_1/graphiques/FigureGraphique.java
@@ -3,13 +3,13 @@ package Revision_Exams.Intra_2.Exercice_1.graphiques;
 import java.awt.Color;
 import java.awt.Graphics;
 
-public class FigureGraphique {
+public abstract class FigureGraphique {
     private final Color couleur;
 
     FigureGraphique(Color couleur) {
         this.couleur = couleur;
     }
 
-    public void dessine(Graphics g) {
+    public abstract void dessine(Graphics g) {
     }
 }
